files ignored formatter setted clean up of the proyect closes #4
